+++ src/structures/mod.rs
@@ -1,4 +1,4 @@
-use std::{f64, iter::zip};
+use std::f64;
